requirements.txt
opencv-python
pyaudio
mysql-connector-python
bcrypt
pillow
numpy
